make configure.sh a bit more efficient

+++ configure.sh
@@ -5,30 +5,30 @@ ONABASE=`cat /etc/onabase`
 type -P netcat &>/dev/null || { echo "Unable to find netcat binary. Please install netcat (http://netcat.sf.net)." >&2; exit 1; }  
 
 # Remove existing symlinks
-rm $ONABASE/bin/cfg_archive
-rm $ONABASE/bin/cfg_archive_procspool
-rm $ONABASE/bin/cfg_archive_cisco_trapscript
-rm $ONABASE/bin/cisco_cfg_expect
-rm $ONABASE/bin/cisco_cmd
-rm $ONABASE/bin/cisco_asa_cmd
-rm $ONABASE/bin/motorola_wing_cmd
-rm $ONABASE/bin/cisco_cfg_snmp
-rm $ONABASE/bin/juniper_cmd
-rm $ONABASE/bin/netscreen_cmd
-rm $ONABASE/bin/extreme_networks_cmd
+# cisco_cmd is not in this list as it is a special case below
+# netscreen_cmd is not in this list as it is a special case below
+BINARIES=( \
+  "cfg_archive" \
+  "cfg_archive_procspool" \
+  "cfg_archive_cisco_trapscript" \
+  "cisco_cfg_expect" \
+  "cisco_asa_cmd" \
+  "motorola_wing_cmd" \
+  "cisco_cfg_snmp" \
+  "juniper_cmd" \
+  "extreme_networks_cmd" \
+)
+for BINARY in "${BINARIES[@]}"
+do
+  [ -f $ONABASE/bin/$BINARY ] && rm $ONABASE/bin/$BINARY
+  ln -s $ONABASE/www/local/plugins/cfg_archive/bin/$BINARY $ONABASE/bin/$BINARY
+done
 
-# Create new symlinks
-ln -s $ONABASE/www/local/plugins/cfg_archive/bin/cfg_archive $ONABASE/bin/cfg_archive
-ln -s $ONABASE/www/local/plugins/cfg_archive/bin/cfg_archive_procspool $ONABASE/bin/cfg_archive_procspool
-ln -s $ONABASE/www/local/plugins/cfg_archive/bin/cfg_archive_cisco_trapscript $ONABASE/bin/cfg_archive_cisco_trapscript
-ln -s $ONABASE/www/local/plugins/cfg_archive/bin/cisco_cfg_expect $ONABASE/bin/cisco_cfg_expect
+# special case binaries
+[ -f $ONABASE/bin/cisco_cmd ] && rm $ONABASE/bin/cisco_cmd
 ln -s $ONABASE/bin/cisco_cfg_expect $ONABASE/bin/cisco_cmd
-ln -s $ONABASE/www/local/plugins/cfg_archive/bin/cisco_asa_cmd $ONABASE/bin/cisco_asa_cmd
-ln -s $ONABASE/www/local/plugins/cfg_archive/bin/motorola_wing_cmd $ONABASE/bin/motorola_wing_cmd
-ln -s $ONABASE/www/local/plugins/cfg_archive/bin/cisco_cfg_snmp $ONABASE/bin/cisco_cfg_snmp
-ln -s $ONABASE/www/local/plugins/cfg_archive/bin/juniper_cmd $ONABASE/bin/juniper_cmd
-ln -s $ONABASE/www/local/plugins/cfg_archive/bin/juniper_cmd $ONABASE/bin/netscreen_cmd
-ln -s $ONABASE/www/local/plugins/cfg_archive/bin/extreme_networks_cmd $ONABASE/bin/extreme_networks_cmd
+[ -f $ONABASE/bin/netscreen_cmd ] && rm $ONABASE/bin/netscreen_cmd
+ln -s $ONABASE/bin/juniper_cmd $ONABASE/bin/netscreen_cmd
 
 # Move our initial configuration files into place
 # Only copy the file over it it does not already exist.
